Collapse the nested-relations wrapper and drop small v1 leftovers
ProcessNestedRelationsHelper had become a pure pass-through around the
orm-v2 implementation (the old v1/v2 branch point), so the real helper
takes over its name and the wrapper is gone; useReplica keeps its false
default. Delete the dead WorkspaceDataSourceService.checkSchemaExists,
make the v2 datasource getRepository generic so the common query runner
no longer needs an as-unknown cast, and replace the structural
RecordQueryBuilder types with the concrete select query builder now
that a single ORM implementation is left.
