+++ packages/twenty-server/src/engine/api/graphql/graphql-query-runner/graphql-query-parsers/utils/apply-filter-entries-to-where-expression.util.ts
@@ -1,10 +1,10 @@
 import { Brackets, NotBrackets, type WhereExpressionBuilder } from 'typeorm';
 
 import { type GraphqlQueryFilterFieldParser } from 'src/engine/api/graphql/graphql-query-runner/graphql-query-parsers/graphql-query-filter/graphql-query-filter-field.parser';
-import { type RecordQueryBuilder } from 'src/engine/api/graphql/graphql-query-runner/types/record-query-builder.type';
+import { type WorkspaceSelectQueryBuilderV2 } from 'src/engine/twenty-orm-v2/query-builder/workspace-select-query-builder-v2';
 
 type FilterWalkContext = {
-  outerQueryBuilder: RecordQueryBuilder;
+  outerQueryBuilder: WorkspaceSelectQueryBuilderV2;
   objectNameSingular: string;
   fieldParser: GraphqlQueryFilterFieldParser;
   useDirectTableReference: boolean;
@@ -19,7 +19,7 @@ export const applyFilterEntriesToWhereExpression = ({
   useDirectTableReference = false,
 }: {
   whereExpression: WhereExpressionBuilder;
-  outerQueryBuilder: RecordQueryBuilder;
+  outerQueryBuilder: WorkspaceSelectQueryBuilderV2;
   objectNameSingular: string;
   filter: Record<string, unknown>;
   fieldParser: GraphqlQueryFilterFieldParser;
